# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sentio SMTP API",
    "description": "Multi-tenant SMTP platform API",
    "contact": {
      "name": "Truespar Team"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/queues/deferred": {
      "get": {
        "tags": [
          "Queues"
        ],
        "operationId": "list_deferred",
        "parameters": [
          {
            "name": "limit",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_Vec_DeferredMessageResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/queues/pause": {
      "post": {
        "tags": [
          "Queues"
        ],
        "operationId": "pause_queue",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_PauseResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/queues/resume": {
      "post": {
        "tags": [
          "Queues"
        ],
        "operationId": "resume_queue",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_PauseResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/queues/stats": {
      "get": {
        "tags": [
          "Queues"
        ],
        "operationId": "queue_stats",
        "parameters": [
          {
            "name": "from",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "path",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_QueueStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/reputation/domains": {
      "get": {
        "tags": [
          "Reputation"
        ],
        "operationId": "list_domain_reputations",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_Vec_DomainReputationSummary"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/reputation/domains/{id}": {
      "get": {
        "tags": [
          "Reputation"
        ],
        "operationId": "get_domain_reputation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_DomainReputationSummary"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/reputation/ips/{ip}": {
      "get": {
        "tags": [
          "Reputation"
        ],
        "operationId": "get_ip_reputation",
        "parameters": [
          {
            "name": "ip",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_IpReputationResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DataResponse_DomainReputationSummary": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "domain_id",
              "domain_name",
              "dmarc_pass",
              "dmarc_fail",
              "dmarc_pass_rate",
              "fbl_complaints"
            ],
            "properties": {
              "dmarc_fail": {
                "type": "integer",
                "format": "int64"
              },
              "dmarc_pass": {
                "type": "integer",
                "format": "int64"
              },
              "dmarc_pass_rate": {
                "type": "number",
                "format": "double"
              },
              "domain_id": {
                "$ref": "#/components/schemas/DomainId"
              },
              "domain_name": {
                "type": "string"
              },
              "fbl_complaints": {
                "type": "integer",
                "format": "int64"
              }
            }
          }
        }
      },
...
      "DataResponse_IpReputationResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "ip",
              "score"
            ],
            "properties": {
              "ip": {
                "type": "string"
              },
              "score": {
                "type": "number",
                "format": "double"
              }
            }
          }
        }
      },
...
      "DataResponse_PauseResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "paused"
            ],
            "properties": {
              "paused": {
                "type": "boolean"
              }
            }
          }
        }
      },
      "DataResponse_QueueStatsResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "counts",
              "paused"
            ],
            "properties": {
              "counts": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/StatusCountResponse"
                }
              },
              "paused": {
                "type": "boolean"
              }
            }
          }
        }
      },
...
      "DataResponse_Vec_DeferredMessageResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "envelope_from",
                "envelope_to",
                "status",
                "created_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "envelope_from": {
                  "type": "string"
                },
                "envelope_to": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "id": {
                  "$ref": "#/components/schemas/MessageId"
                },
                "status": {
                  "$ref": "#/components/schemas/MessageStatus"
                },
                "subject": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              }
            }
          }
        }
      },
...
      "DataResponse_Vec_DomainReputationSummary": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "domain_id",
                "domain_name",
                "dmarc_pass",
                "dmarc_fail",
                "dmarc_pass_rate",
                "fbl_complaints"
              ],
              "properties": {
                "dmarc_fail": {
                  "type": "integer",
                  "format": "int64"
                },
                "dmarc_pass": {
                  "type": "integer",
                  "format": "int64"
                },
                "dmarc_pass_rate": {
                  "type": "number",
                  "format": "double"
                },
                "domain_id": {
                  "$ref": "#/components/schemas/DomainId"
                },
                "domain_name": {
                  "type": "string"
                },
                "fbl_complaints": {
                  "type": "integer",
                  "format": "int64"
                }
              }
            }
          }
        }
      },
...
      "DomainId": {
        "type": "string",
        "format": "uuid"
      },
...
      "ErrorBody": {
        "type": "object",
        "required": [
          "type",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        }
      },
...
      "MessageId": {
        "type": "string",
        "format": "uuid"
      },
...
      "MessageStatus": {
        "type": "string",
        "enum": [
          "queued",
          "processing",
          "delivered",
          "deferred",
          "bounced",
          "dropped",
          "scheduled",
          "held"
        ]
      },
...
      "StatusCountResponse": {
        "type": "object",
        "required": [
          "status",
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    }
  },
...
}